+++ corelibs/trunk/render.immed/src/cytoscape/render/immed/nodeshape/DiamondNodeShape.java
@@ -49,14 +49,14 @@ public class DiamondNodeShape extends AbstractNodeShape {
 		path = new GeneralPath(); 
 	}
 		
-	public Shape getShape(double xMin, double yMin, double xMax, double yMax) {
+	public Shape getShape(float xMin, float yMin, float xMax, float yMax) {
 
 		path.reset();
 
-		path.moveTo((xMin + xMax) / 2.0d, yMin);
-		path.lineTo(xMax, (yMin + yMax) / 2.0d);
-		path.lineTo((xMin + xMax) / 2.0d, yMax);
-		path.lineTo(xMin, (yMin + yMax) / 2.0d);
+		path.moveTo((xMin + xMax) / 2.0f, yMin);
+		path.lineTo(xMax, (yMin + yMax) / 2.0f);
+		path.lineTo((xMin + xMax) / 2.0f, yMax);
+		path.lineTo(xMin, (yMin + yMax) / 2.0f);
 
 		path.closePath();
