+++ corelibs/trunk/render.immed/src/cytoscape/render/immed/nodeshape/HexagonNodeShape.java
@@ -49,16 +49,16 @@ public class HexagonNodeShape extends AbstractNodeShape {
 		path = new GeneralPath(); 
 	}
 		
-	public Shape getShape(double xMin, double yMin, double xMax, double yMax) {
+	public Shape getShape(float xMin, float yMin, float xMax, float yMax) {
 
 		path.reset();
 
-		path.moveTo(((2.0d * xMin) + xMax) / 3.0d, yMin);
-		path.lineTo(((2.0d * xMax) + xMin) / 3.0d, yMin);
-		path.lineTo(xMax, (yMin + yMax) / 2.0d);
-		path.lineTo(((2.0d * xMax) + xMin) / 3.0d, yMax);
-		path.lineTo(((2.0d * xMin) + xMax) / 3.0d, yMax);
-		path.lineTo(xMin, (yMin + yMax) / 2.0d);
+		path.moveTo(((2.0f * xMin) + xMax) / 3.0f, yMin);
+		path.lineTo(((2.0f * xMax) + xMin) / 3.0f, yMin);
+		path.lineTo(xMax, (yMin + yMax) / 2.0f);
+		path.lineTo(((2.0f * xMax) + xMin) / 3.0f, yMax);
+		path.lineTo(((2.0f * xMin) + xMax) / 3.0f, yMax);
+		path.lineTo(xMin, (yMin + yMax) / 2.0f);
 
 		path.closePath();
