+++ corelibs/trunk/render.immed/src/cytoscape/render/immed/nodeshape/TriangleNodeShape.java
@@ -49,11 +49,11 @@ public class TriangleNodeShape extends AbstractNodeShape {
 		path = new GeneralPath(); 
 	}
 		
-	public Shape getShape(double xMin, double yMin, double xMax, double yMax) {
+	public Shape getShape(float xMin, float yMin, float xMax, float yMax) {
 
 		path.reset();
 
-		path.moveTo((xMin + xMax) / 2.0d, yMin);
+		path.moveTo((xMin + xMax) / 2.0f, yMin);
 		path.lineTo(xMax, yMax);
 		path.lineTo(xMin, yMax);
